Fix pre7.4 again
